Set a unique user agent string for the exporter
We provide paas-prometheus-exporter as a bit of software that tenants can use,
and use it they do. Unfortunately, depending on the credentials it is given, it
can end up making a lot of API calls, and get itself rate limited.

As operators, it would be nice if we were able to correlate the HTTP 429
responses we see as a result of rate limiting with a particular (instance of a)
paas-prometheus-exporter deployment. This commit adds a unique user agent
string to the CF API requests which includes the app id and app index, so that
we can do just that.